Clean the Code, added JavaDoc, modified types and useless commands

+++ src/drivers/PageRankSchimmyRun.java
@@ -7,6 +7,7 @@ import org.apache.hadoop.conf.Configured;
 import org.apache.hadoop.fs.Path;
 import org.apache.hadoop.io.DoubleWritable;
 import org.apache.hadoop.io.IntWritable;
+import org.apache.hadoop.io.LongWritable;
 import org.apache.hadoop.mapreduce.Job;
 import org.apache.hadoop.mapreduce.lib.input.FileInputFormat;
 import org.apache.hadoop.mapreduce.lib.output.FileOutputFormat;
@@ -15,6 +16,12 @@ import org.apache.hadoop.util.Tool;
 import partitioners.SchimmyPartitioner;
 import reducers.PageRankSchimmyReducer;
 
+/**
+ * Classe driver per l'avvio del job riguardante il calcolo del pagerank con Schimmy design pattern.
+ * 
+ * @author Nicolò Marchi, Fabio Pettenuzzo
+ *
+ */
 public class PageRankSchimmyRun extends Configured implements Tool {
 
 	@Override
@@ -27,24 +34,17 @@ public class PageRankSchimmyRun extends Configured implements Tool {
 		job.setJarByClass(PageRankSchimmyRun.class);
 		job.setMapperClass(PageRankSchimmyMapper.class);
 		
-		job.setMapOutputKeyClass(IntWritable.class);
+		job.setMapOutputKeyClass(LongWritable.class);
 		job.setMapOutputValueClass(DoubleWritable.class);
 		
 		job.setPartitionerClass(SchimmyPartitioner.class);
-//		job.setPartitionerClass(TotalOrderPartitioner.class);
 
-//		job.setSortComparatorClass(IntWritable.Comparator.class);
-		
 		job.setReducerClass(PageRankSchimmyReducer.class);
 		job.setNumReduceTasks(Integer.parseInt(args[2]));
 
 		FileInputFormat.addInputPath(job, new Path(args[0]));
 		FileOutputFormat.setOutputPath(job, new Path(args[1]));
-//		
-//		MultipleOutputs.addNamedOutput(job, "log", TextOutputFormat.class, Text.class, Text.class);
-//
-//	    MultipleOutputs.addNamedOutput(job, "output", TextOutputFormat.class, Text.class, Text.class);
-		
+
 		return job.waitForCompletion(true) ? 0 : 1;
 
 	}
